Add files via upload
1.Given a number n, find the sum of first natural numbers.
Examples: 

Input : n = 3
Output : 6
Explanation :
Note that 1 + 2 + 3 = 6

Input  : 5
Output : 15 
Explanation :
Note that 1 + 2 + 3 + 4 + 5 = 15
A simple solution is to do the following. 

1) Initialize : sum = 0
2) Run a loop from x = 1 to n and 
   do following in loop.
     sum = sum + x 

2. Count Digits

In a realm where numbers hold secrets, a captivating challenge awaits, which is to, Count digits in a number !!!

Our Task: We are given a number with n number of digits, and our task is to return the count of digits in this number

Examples:

Input: 256
Output: 3
 
Input: 75
Output: 2
 
Input: 5
Output: 1
Simple Iterative Solution to count digits in an integer
Steps Followed:

The integer entered by the user is stored in the variable n. 
Then the while loop is iterated until the test expression n != 0 is evaluated to 0 (false).  
 Let us consider 3456 as the input integer.  

After the first iteration, the value of n will be updated to 345 and the count is incremented to 1.
After the second iteration, the value of n will be updated to 34 and the count is incremented to 2.
After the third iteration, the value of n will be updated to 3 and the count is incremented to 3.
In the fourth iteration, the value of n will be updated to zero and the count will be incremented to 4.  
Then the test expression is evaluated ( n!=0 ) as false and the loop terminates with final count as 4.