Add files via upload
Our Task: We are given an n-digit number. The task is to find if the number is palindrome or not, provided that, n>=0.

 

What is a Palindrome Number?

If the Reversed Number is equal to the original number, then the number is said to be a palindrome number, otherwise not.

Examples:

Input: 78987Output: YesInput: 8668Output: Yes Input: 8Output: Yes Input: 21Output: No Input: 837Output: No
 

Simple approach to check if a number is Palindrome or not:
Steps followed:

Let us consider the input number to be 3443

Declare a variable rev and the reversed number will get stored in it as the iteration continues

After the first iteration, the value of n will be updated to 344, and rev = 3.
After the second iteration, the value of n will be updated to 34, and rev =34
After the third iteration, the value of n will be updated to 3, and rev =344
In the fourth iteration, the value of n will be updated to zero, and rev =3443
Then the test expression is evaluated ( n!=0 ) as false and the loop terminates.
Finally, compare the reverse number to the original number. If it is the same, then the number is palindrome otherwise not.
Time Complexity: Theta(d), where d is the number of digits present in the number.

Auxiliary space : O(1) or constant



Our Task: We are given a number. The task is to find the factorial of the number.

What is the factorial of a number?
Factorial of a non-negative integer is the multiplication of all positive integers smaller than or equal to n. 
For example, the factorial of 6 is 6*5*4*3*2*1 which is 720. 
A factorial is represented by a number and a  ” ! ”  mark at the end. It is widely used in permutations and combinations to calculate the total possible outcomes. A French mathematician Christian Kramp first used the exclamation.
The factorial (n!) of a number is the product of all the integers from 1 to that number, i.e, 

                                                     n! = 1*2*3*.....(n-1)*n
 Examples:

Input: 4 Output: 4*3*2*1 = 24
Input: 6 Output: 6*5*4*3*2*1 = 720
Input: 0 Output: 1 
Input: 1 Output: 1
 

We have 2 approaches to solve the problem: Iterative approach & Recursive approach

1) Iterative Approach:
Steps:

Using a for loop, we will write a program for finding the factorial of a number. 
An integer variable with a value of 1 will be used in the program. 
With each iteration, the value will increase by 1 until it equals the value entered by the user. 
The factorial of the number entered by the user will be the final value in the fact variable.

Time Complexity: O(n)
Auxiliary Space: O(1)

2) Recursive Approach

Time Complexity: O(n)

Space Complexity: O(n) // Recursion requires stack space to store function calls
A recursive function is used to calculate the factorial, with a base condition to return 1, when n=0